// src/apis/models/FireBytes_Unified_Wave_ControlPanels_Dto_DevicePointViewWithNavigationPropertiesDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_ControlPanels_Dto_AmsBuildingFloorDto } from './FireBytes_Unified_Wave_ControlPanels_Dto_AmsBuildingFloorDto';
import type { FireBytes_Unified_Wave_ControlPanels_Dto_AmsDeviceDto } from './FireBytes_Unified_Wave_ControlPanels_Dto_AmsDeviceDto';
import type { FireBytes_Unified_Wave_ControlPanels_Dto_AmsDevicePointViewDto } from './FireBytes_Unified_Wave_ControlPanels_Dto_AmsDevicePointViewDto';
export type FireBytes_Unified_Wave_ControlPanels_Dto_DevicePointViewWithNavigationPropertiesDto = {
    amsDevicePointView?: FireBytes_Unified_Wave_ControlPanels_Dto_AmsDevicePointViewDto;
    amsDevice?: FireBytes_Unified_Wave_ControlPanels_Dto_AmsDeviceDto;
    amsBuildingFloor?: FireBytes_Unified_Wave_ControlPanels_Dto_AmsBuildingFloorDto;
};