// src/apis/models/FireBytes_Unified_Wave_ControlPanels_Dto_DeviceStatusDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanels_Dto_DeviceStatusDto = {
    id?: string;
    statusName?: string | null;
    statusIconFont?: string;
    statusCode?: string | null;
    description?: string | null;
    messageBackgroundColorCode?: string | null;
    deviceType?: string | null;
    deviceTypeId?: string | null;
};